+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### 2.4.10 (TBD)
+### 2.4.10 (January 13, 2022)
 
 - Feature: The flag `--http-plaintext` can be used to ensure that an intercept uses plaintext http or grpc when 
   communicating with the workstation process.